Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/KeyframeEffect/getKeyframes-expected.txt


FAIL getKeyframes() should serialize its css values with a on keyframe sequence which requires value serilaization of its values assert_equals: value for 'backgroundColor' on ComputedKeyframe #0 expected "rgb(1, 2, 3)" but got "rgb(1,2,3)"
